Update README.md
Atliq Grands noticed a loss in their market share and revenue over a few months. To understand the cause
of this loss, they needed a way to analyse this. I created a dashboard in Power BI using three months of
data.
With the created dashboard, Revenue team of Atliq Grands were able to gain insights about their revenue
trend. This could help in regaining their revenue and market share by 20% in the next month.